separate file-view into reusable component for use with by result display component
